Update PlayerMovement class to use Run input
The PlayerMovement MonoBehaviour now uses its IInputProvider to check if
the player wants to run and adjusts the movement speed accordingly.

The target movement speed is multiplied by a factor set in the Editor if
the player wants to run, and the value is changed smoothly using Mathf's
SmoothDamp method.